Fix support for Watchify
Currently jadeify only registers the required jade file, not allowing watchify to rebuild or cache changes to the jade dependency tree.  This modification uses the Jade parser to create an array of the dependency tree of a required jade file and registers its dependencies as "files" for browserify and watchify.